snippets-app: fix except block to rollback connection.rollback()
Previously was removed on accident. See Unit 2 Lesson 1.5, Coping
with duplicates section. The with statement, which is a context
manager will automatically perform cleanup actions when the with
block is exited. But it appears that it does not do that in middle
of the statement.

When the try block fails with the insert, the connection needs to be
rolled back in the except block before performing the update command.